added more data when posting a listing

+++ app/src/main/java/com/shoppr/app/data/user/model/User.java
@@ -8,8 +8,8 @@ public class User {
     private String email;
     private String phoneNumber;
 	private String address;
-    private long latitude;
-    private long longitude;
+    private double latitude;
+    private double longitude;
 
 	public User() {
 	}
@@ -82,19 +82,19 @@ public class User {
         this.address = address;
     }
 
-    public long getLatitude() {
+    public double getLatitude() {
         return latitude;
     }
 
-    public void setLatitude(long latitude) {
+    public void setLatitude(Double latitude) {
         this.latitude = latitude;
     }
 
-    public long getLongitude() {
+    public double getLongitude() {
         return longitude;
     }
 
-    public void setLongitude(long longitude) {
+    public void setLongitude(Double longitude) {
         this.longitude = longitude;
     }
 }
